InteriorPurchase

+++ Assets/CyberClub/Scripts/Resources/ResourcesData/IResource.cs
@@ -4,5 +4,5 @@ public interface IResource
 {
     ResourceType Type {get; set;}
     bool TryBuy(int amount);
-    void AddResource(int amount, int multiplier);
+    void AddResource(int amount, float multiplier);
 }
